Add benchmarking fuzzer component
** Last Driver component is a fuzzer that will start execution using
** the corpus generated by river.genetic. This component is a
** substitute for a future runtime coverage evaluator.

+++ scripts/run.sh
@@ -1,30 +1,41 @@
 #!/usr/bin/env bash
 
-[ $# -lt 6 ] && { echo "Usage: $0 [binary-id] [runs] [cores] [benchmark-runs] [(no)rebuild] [(no)regenerate-corpus]"; exit 1; }
-read binary_id runs cores benchmark_runs rebuild regenerate_corpus <<<$@
+[ $# -lt 6 ] && { echo "Usage: $0 [binary-id] [runs] [cores] [benchmark-runs] [(no)rebuild] [(no)regenerate-corpus] [genetic]"; exit 1; }
+read binary_id runs cores benchmark_runs rebuild regenerate_corpus genetic<<<$@
 
 RUNS=$runs
 DRIVER_DIR=$(pwd)
-RIVER_GENETIC=$DRIVER_DIR/river.genetic
+GRIVER=$DRIVER_DIR/river.genetic
 TRACER_NODE=$DRIVER_DIR/tracer.node
 NODE_RIVER=$TRACER_NODE/deps/node-river
 
 CONFIG_PATH=$(readlink -f config.json)
 
 DB_NAME="tests_$binary_id"
-LOGS_DIR=$(pwd)/logs/$binary_id
+
+ID_LOGS_DIR=$(pwd)/$binary_id
+LOGS_DIR=$ID_LOGS_DIR/logs
+RESULTS_DIR=$ID_LOGS_DIR/results
+
 STATS_FILE=$LOGS_DIR/stats-$cores.csv
 FUZZER_STATS_FILE=$LOGS_DIR/fuzzer-stats-$cores.csv
+GRIVER_STATS=$LOGS_DIR/river-genetic-fuzzer-$cores.csv
+BASELINE=$LOGS_DIR/baseline-$cores.csv
 LOG_FILE=$LOGS_DIR/run-$cores.log
 FUZZER_LOG_FILE=$LOGS_DIR/fuzzer-$cores.log
-GENETIC_LOG_FILE=$LOGS_DIR/river-genetic-$cores.log
+DEFUZZER_LOG_FILE=$LOGS_DIR/defuzzer-$cores.log
+GRIVER_LOG_FILE=$LOGS_DIR/river-genetic-$cores.log
+
 NO_TESTCASES=0
 
 cleanup() {
   # clean the DRIVER build and stop node
   echo -e "\033[0;32m[DRIVER] Cleaning DRIVER environment ..."; echo -e "\033[0m"
   killall -9 node
-  killall -9 python3
+  killall -9 fuzzer
+  if [ "$genetic" == "genetic" ]; then
+    killall -9 python3
+  fi
 
   if [ "$rebuild" == "rebuild" ]; then
     cd $TRACER_NODE && rm -rf build && npm install
@@ -46,6 +57,23 @@ cleanup() {
   # purge rabbit queues
   sudo rabbitmqctl purge_queue driver.newtests.$binary_id
   sudo rabbitmqctl purge_queue driver.tracedtests.$binary_id
+
+  # remove results dumped to disk
+  if [ -d $RESULTS_DIR ]; then
+    rm -rf $RESULTS_DIR
+  fi
+}
+
+sigint_handler()
+{
+  echo -e "\033[0;32m[DRIVER] Received SIGINT. Cleaning DRIVER environment ..."; echo -e "\033[0m"
+
+  sudo systemctl stop mongo.rabbit.bridge
+  killall -9 node
+  killall -9 fuzzer
+  killall -9 python3
+
+  exit 1
 }
 
 generate_testcases() {
@@ -66,7 +94,7 @@ generate_testcases() {
 
   export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib
   rm -f $FUZZER_LOG_FILE
-  $fuzzer_path -close_fd_mask=3 -initial_corpus_db=1 \
+  $fuzzer_path -close_fd_mask=3 -initial_corpus_db=1 -dump_all_testcases=1 \
      -dump_to_db=1 -config=$CONFIG_PATH \
     -binary_id=$binary_id -print_final_stats=1 >> $FUZZER_LOG_FILE 2>&1 &
 
@@ -85,7 +113,7 @@ start_tracer() {
   fuzzer_path=$(pwd)/$binary_id/fuzzer
 }
 
-river_genetic_environment() {
+griver_environment() {
   ## genetic alg setup
   ## TODO get rid of this stuff
   export SPARK_HOME=/data/spark/spark-2.0.1-bin-hadoop2.7
@@ -99,31 +127,38 @@ river_genetic_environment() {
   export SPARK_LOCAL_IP=10.18.0.32
 
   export PYTHONUNBUFFERED=1
-  export SIMPLETRACERLOGSPATH=$RIVER_GENETIC/genetic_code/logs
+  export SIMPLETRACERLOGSPATH=$RESULTS_DIR
   #export LD_LIBRARY_PATH=/data/simpletracer/build-river-tools/lib
   #export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/data/simpletracer/river.sdk/lin/lib
   #export SIMPLETRACERPATH=/data/simpletracer/build-river-tools/bin/simple_tracer
 }
 
-start_river_genetic() {
+start_griver() {
   echo -e "\033[0;32m[DRIVER] Starting river genetic for $binary_id executable..."; echo -e "\033[0m"
-  cd $RIVER_GENETIC
-  river_genetic_environment
+  cd $GRIVER
+  griver_environment
   ( python3 ./genetic_code/main.py --testsFolderCount 10 --numTasks 1 \
       --isParallelExecution 0 --populationSize 10 --numberOfIterations 10 \
       --config $CONFIG_PATH --driver 1  --isDebugEnabled 1 \
-      --executableId $binary_id > $GENETIC_LOG_FILE 2>&1 &)
+      --executableId $binary_id > $GRIVER_LOG_FILE 2>&1 &)
 }
 
 wait_for_termination() {
   while true; do
-    left=$(mongo --eval "db.$DB_NAME.find({state : \"traced\"}).count()" | tail -n1)
+    left=$( mongo --eval "db.$DB_NAME.find({state : \"traced\"}).count()" 2> /dev/null | tail -n1; \
+      exitcode=${PIPESTATUS[0]}; if test $exitcode -ne 0; then echo -n -1; fi )
+    if [ $left -lt 0 ]; then
+      echo -e "\033[0;32m[DRIVER] Mongo could not connect. Exiting...."; echo -e "\033[0m"
+      sleep 30
+      break
+    fi
     sleep 3 #todo fix this
     echo "Tracing progress: found [$left] testcases traced."
   done
 }
 
 print_statistics() {
+  sleep 10
   fst=$(mongo --eval "db.$DB_NAME.find({}, {tracedTs:1, _id:0}).sort({tracedTs: 1}).limit(1)" | tail -n 1 | grep --only-matching --perl-regex "(?<=\"tracedTs\" : )[^ ]*")
   lst=$(mongo --eval "db.$DB_NAME.find({}, {tracedTs:1, _id:0}).sort({tracedTs: -1}).limit(1)" | tail -n 1 | grep --only-matching --perl-regex "(?<=\"tracedTs\" : )[^ ]*")
   printf "%30s\t%12s\t%8s\t%20s\n" "$binary_id" "$NO_TESTCASES" "$cores" "$((lst - fst))" >> $STATS_FILE
@@ -135,6 +170,31 @@ print_statistics() {
 
   printf "%30s\t%12s\t%8s\t%8s\t%8s\t%8s\n" "$binary_id" "$NO_TESTCASES" "$cores" "$execs" "$execs_per_sec" "$units" >> $FUZZER_STATS_FILE
 
+}
+
+evaluate_new_corpus() {
+  ## Driver is a system that uses testcases in order to generate new ones.
+  ## First set of testcases is generated by a raw fuzzer.
+  ## These testcases are traced by river and all traces go in mongo.
+  ## A special component will retrieve all traced testcases from mongo
+  ##  and will generate new testcases based on the distribution of
+  ##  cold spots and hot spots in traced testcases.
+  ##
+  ## This function evaluates the new corpus generated in $RESULTS_DIR
+  ## The result is a set of (testcaseId, coverage) dumped in csv
+
+  ## run all testcases and do not add anything new
+  $fuzzer_path $RESULTS_DIR -close_fd_mask=3 -runs=$RUNS \
+    -print_final_stats=1 >> $DEFUZZER_LOG_FILE 2>&1
+
+  ### 131072pulse  cov: 53 ft: 83 corp: 51/1011b exec/s: 65536 rss:  40Mb
+  pulsecov="$(grep -o '#[0-9]*[^p]*[pulse\|NEW] *cov: [0-9]*' $DEFUZZER_LOG_FILE | grep -P  '(\d+)' -o)"
+  ### pulsecov is like (id\ncov\nid\ncov ...)
+  printf "%20s\t%8s\n" $pulsecov >> $GRIVER_STATS
+
+  pc="$(grep -o '#[0-9]*[^p]*[pulse\|NEW] *cov: [0-9]*' $FUZZER_LOG_FILE | grep -P  '(\d+)' -o)"
+  ### pulsecov is like (id\ncov\nid\ncov ...)
+  printf "%20s\t%8s\n" $pc >> $BASELINE
 }
 
 main() {
@@ -158,15 +218,21 @@ main() {
   echo "======================================================="
   printf "%30s\t%12s\t%8s\t%20s\n" "name" "testcases" "cores" "time_in_ms" >> $STATS_FILE
   printf "%30s\t%12s\t%8s\t%8s\t%8s\t%8s\n" "name" "testcases" "cores" "execs" "execs_per_sec" "units" >> $FUZZER_STATS_FILE
+  printf "%20s\t%8s\n" "testcaseid" "cov" >> $GRIVER_STATS
+  printf "%20s\t%8s\n" "testcaseid" "cov" >> $BASELINE
 
   for i in $(seq $benchmark_runs); do
     cleanup
     start_tracer
-    start_river_genetic
+    if [ "$genetic" == "genetic" ]; then
+      start_griver
+    fi
     generate_testcases
     wait_for_termination
     print_statistics
+    evaluate_new_corpus
   done
 }
 
+trap sigint_handler SIGINT
 main
